refactor: move statistic charts state to contexts
